+++ java/BV1uJ411k7wy/AVariables/Demo05DataType.java
@@ -1,4 +1,4 @@
-package Day0001;
+package AVariables;
 
 /**
  * 强制类型转换
