+++ java/BV1uJ411k7wy/BFunction/Demo03MethodParam.java
@@ -1,4 +1,4 @@
-package Day0002;
+package BFunction;
 
 public class Demo03MethodParam {
     public static void main(String[] args) {
