+++ java/BV1uJ411k7wy/BFunction/Demo04MethodReturn.java
@@ -1,4 +1,4 @@
-package Day0002;
+package BFunction;
 
 public class Demo04MethodReturn {
     public static void main(String[] args) {
