+++ java/BV1uJ411k7wy/BFunction/Demo05MethodSame.java
@@ -1,4 +1,4 @@
-package Day0002;
+package BFunction;
 
 public class Demo05MethodSame {
     public static void main(String[] args) {
